requirements.txt
Django==4.2.7
Pillow==10.1.0
rembg==2.0.50
opencv-python==4.8.1.78
numpy==1.24.3
